requirements.txt
Appium-Python-Client==1.0.2
atomicwrites==1.4.1
attrs==22.2.0
colorama==0.4.6
iniconfig==1.1.1
packaging==22.0
pluggy==0.13.1
py==1.11.0
pytest==6.1.1
selenium==3.141.0
toml==0.10.2
urllib3==1.26.13
